:recycle: Move If-Extension to own file
As we are going to need this file in multiple views,
its better to separate it from
MergeConflictResolutionView.
